+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Color tags to VTML
 - Argument parsing group for table formats.
 - Table is a context manager that calls close when exited.
+- CSV, JSON and Markdown table renderers.
 
 ### Changed
 - VTML rendering of --help output;  Emboldened some elements.